+++ src/components/themed/WalletListCreateRow.tsx
@@ -32,9 +32,9 @@ export interface WalletListCreateRowProps {
   onPress?: (walletId: string, tokenId: EdgeTokenId) => Promise<void>
 }
 
-export const WalletListCreateRowComponent = (
-  props: WalletListCreateRowProps
-): React.ReactElement => {
+export const WalletListCreateRowComponent: React.FC<
+  WalletListCreateRowProps
+> = props => {
   const {
     createItem,
     createWalletId,
@@ -160,7 +160,11 @@ export const WalletListCreateRowComponent = (
   })
 
   return (
-    <EdgeTouchableOpacity style={styles.row} onPress={handlePress}>
+    <EdgeTouchableOpacity
+      style={styles.row}
+      testID={`walletListCreateRow_${pluginId ?? 'unknown'}`}
+      onPress={handlePress}
+    >
       <CryptoIcon
         marginRem={1}
         pluginId={pluginId}