+++ src/locales/en_US.ts
@@ -748,6 +748,7 @@ const strings = {
   string_first_dash_wallet_name: 'My Dash',
   string_first_digibyte_wallet_name: 'My DigiByte',
   string_first_ecash_wallet_name: 'My eCash',
+  string_first_ecashcom_wallet_name: 'My eCash.com',
   string_first_eos_wallet_name: 'My EOS',
   string_first_holesky_wallet_name: 'My Holesky',
   string_first_hyperevm_wallet_name: 'My HyperEVM',